chore: move render method to separate file
