+++ elements/chart/src/methods/render/index.js
@@ -1,0 +1,1 @@
+export { default as renderChartMethod } from "./render-chart";
